requirement.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.8
Faker==33.1.0
inflection==0.5.1
kombu==5.4.2
packaging==24.2
prompt_toolkit==3.0.48
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rabbitmq==0.2.0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
